Changes to demonstrate Field Injection vs Constructor Injection in Spring
