+++ spring-all/src/main/java/org/baeldung/user/controller/UserController.java
@@ -1,8 +1,10 @@
 package org.baeldung.user.controller;
 
-import org.baeldung.persistence.model.MyUser;
+import org.baeldung.user.model.MyUser;
+import org.baeldung.user.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -12,6 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserController {
 
     public static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
+    
+    @Autowired
+    private UserService userService;
 
     /**
      * Controller method to fetch user object based on user name
@@ -22,6 +27,6 @@ public class UserController {
     @RequestMapping("/user")
     @ResponseBody
     public MyUser getUser(@RequestParam String userName) {
-        return null;
+        return userService.getUserByUsername(userName);
     }
 }
