+++ spring-all/src/main/java/org/baeldung/user/model/MyUser.java
@@ -1,4 +1,4 @@
-package org.baeldung.persistence.model;
+package org.baeldung.user.model;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
